third_party/blink/web_tests/svg/dynamic-updates/SVGFEDistantLightElement-svgdom-azimuth-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'azimuth' property of the SVGFEDistantLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS distantLight.azimuth.baseVal is 10
PASS distantLight.azimuth.baseVal is 100
PASS successfullyParsed is true

TEST COMPLETE
